pyproject.toml
[project]
name = "auto-prompt-optimizer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "python-dotenv>=1.2.2",
]
